+++ system/src/Grav/Framework/Psr7/UploadedFile.php
@@ -23,6 +23,9 @@ class UploadedFile implements UploadedFileInterface
 {
     use UploadedFileDecoratorTrait;
 
+    /** @var array */
+    private $meta = [];
+
     /**
      * @param StreamInterface|string|resource $streamOrFile
      * @param int                             $size
@@ -33,5 +36,35 @@ class UploadedFile implements UploadedFileInterface
     public function __construct($streamOrFile, $size, $errorStatus, $clientFilename = null, $clientMediaType = null)
     {
         $this->uploadedFile = new \Nyholm\Psr7\UploadedFile($streamOrFile, $size, $errorStatus, $clientFilename, $clientMediaType);
+    }
+
+    /**
+     * @param array $meta
+     * @return $this
+     */
+    public function setMeta(array $meta)
+    {
+        $this->meta = $meta;
+
+        return $this;
+    }
+
+    /**
+     * @param array $meta
+     * @return $this
+     */
+    public function addMeta(array $meta)
+    {
+        $this->meta = array_merge($this->meta, $meta);
+
+        return $this;
+    }
+
+    /**
+     * @return array
+     */
+    public function getMeta(): array
+    {
+        return $this->meta;
     }
 }